Handle SIGTERM for clean shutdown
Fixes #6.

Prior to this change, a DBus signal from the SessionManager is the only
way for Cerbere to initiate shutdown. If the connection to DBus is
lost, Cerbere will keep running until externally killed (e.g. by
systemd at shutdown time after its configured timeout).

Since systemd sends SIGTERM to running processes in user sessions
before resorting to SIGHUP, we can handle it much the same way as if
the SessionManager was asking Cerbere to exit and behave correctly.